requirement.txt
Flask==2.3.3
flask-cors==4.0.0
pdf2docx==0.5.6
docx2pdf==0.1.8
pdf2image==1.16.3
Pillow==10.0.1
pytesseract==0.3.10
python-docx==0.8.11
PyMuPDF==1.23.8
Werkzeug==2.3.7